requirements.txt
fastapi
fastapi-mcp
faiss-cpu
uvicorn[standard]
python-dotenv
pydantic
torch
clip @ git+https://github.com/openai/CLIP.git
easyocr
Pillow
pdf2image
qdrant-client
numpy
python-multipart
werkzeug
requests
PyYAML
google-adk
google-generativeai
pymupdf
nest_asyncio
langchain_text_splitters
watchdog
pinecone-text
mcp